refactor: extract display module to top level
