+++ examples/GPU/example_3d_trajectory_display.py
@@ -17,7 +17,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 import mrinufft.trajectories.inits as mtt
-from mrinufft.trajectories.display import get_gridded_trajectory
+from mrinufft.display import get_gridded_trajectory
 from mrinufft.trajectories.utils import Acquisition
 
 BACKEND = os.environ.get("MRINUFFT_BACKEND", "cufinufft")
