+++ src/mrinufft/display/__init__.py
@@ -1,0 +1,15 @@
+"""Display functions for trajectories."""
+
+from .advanced import get_gridded_trajectory
+from .config import displayConfig
+from .gradients import display_gradients, display_gradients_simply
+from .trajectories import display_2D_trajectory, display_3D_trajectory
+
+__all__ = [
+    "displayConfig",
+    "display_2D_trajectory",
+    "display_3D_trajectory",
+    "display_gradients",
+    "display_gradients_simply",
+    "get_gridded_trajectory",
+]
