+++ src/mrinufft/display/trajectories.py
@@ -8,7 +8,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 from numpy.typing import NDArray
 
-from mrinufft.trajectories.display.config import displayConfig
+from .config import displayConfig
 from mrinufft.trajectories.utils import (
     Acquisition,
     compute_gradients_and_slew_rates,
