--- src/mrinufft/trajectories/display3D.py
@@ -1,6 +1,0 @@
-"""Backward-compatibility shim for display3D functions.
-
-Advanced 3D display utilities have moved to the display/ submodule.
-"""
-
-from mrinufft.trajectories.display.advanced import get_gridded_trajectory
